# software-engineering-2-backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: "myWorld Travel API"
  description: |
    ## 🔗 **HATEOAS Architecture (REST Level 3)**
    This API implements **Hypermedia As The Engine Of Application State (HATEOAS)** following REST Level 3 architectural constraints:
    - **Dynamic Navigation**: Every response includes discoverable links to related resources and available actions
    - **Self-Descriptive**: Clients can navigate the API entirely through hypermedia links without hardcoding URLs
    - **Stateless Interaction**: All possible actions are dynamically provided based on current resource state

    Each resource response includes a `links` object containing:
    - Available HTTP methods and URIs for resource operations
    - Navigation paths to connected entities  
    - Self-reference links to the current resource

    **Important**: Link availability is dynamic and depends on the current resource state, user permissions, and context. The examples show typical links but actual responses may include additional or fewer links based on the specific situation.

    ---
    🌍 **myWorld Travel API** - Your intelligent travel companion for discovering the world!

    A comprehensive RESTful API that powers personalized travel experiences through smart recommendations and community-driven insights.

    ### 🚀 **Key Features**
    - **AI-Powered Recommendations**: Intelligent place suggestions based on your preferences, location, and travel history
...
    
    ### 👤 **User Management**
...

...

...

...

...

...
servers:
...
paths:
...
  "/users/{userId}/disliked-places":
    get:
      summary: "Get user's disliked places"
      operationId: getDislikedPlaces
      description: |
        **FR24 - The user should be able to view places excluded from their recommendations.**

        Retrieves all places that the user has marked as disliked and are excluded from recommendations.
      tags: ["Favourite & Disliked Places"]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          description: "Unique identifier of the user whose disliked places are being retrieved"
      responses:
        '200':
          description: "Successfully retrieved list of all disliked places for the user"
          content:
            application/json:
              schema:
                type: object
                properties:
                  dislikedPlaces:
                    type: array
                    items:
                      allOf:
                        - $ref: "#/components/schemas/DislikedPlace"
                        - type: object
                          properties:
                            links:
                              type: object
                              additionalProperties:
                                $ref: "#/components/schemas/HypermediaLink"
                              description: "Available actions for this disliked place"
                              example:
                                select:
                                  href: "/places/4321"
                                  method: GET
                                remove-disliked:
                                  href: "/users/16180/disliked-places/456"
                                  method: DELETE
                  links:
                    type: object
                    additionalProperties:
                      $ref: "#/components/schemas/HypermediaLink"
                    description: "Available actions for disliked places collection"
                    example:
                      self:
                        href: "/users/16180/disliked-places"
                        method: GET
                      add-disliked:
                        href: "/users/16180/disliked-places"
                        method: POST
                      user-profile:
                        href: "/users/16180/profile"
                        method: GET
        '404':
          description: "User not found - No user exists with the specified userId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: USER_NOT_FOUND
                message: "User with ID 12345 not found"
    post:
      summary: "Add a place to user's disliked places"
      operationId: addDislikedPlace
      description: |
        **FR24 - The user should be able to exclude a place from their recommendations.**

        Adds a place to the user's disliked places to exclude it from future recommendations.
      tags: ["Favourite & Disliked Places"]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          description: "Unique identifier of the user adding a place to their dislikes for recommendation exclusion"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [placeId]
              properties:
                placeId:
                  type: integer
                  description: "Unique identifier of the place to be excluded from future recommendations"
      responses:
        '201':
          description: "Place successfully added to user's disliked places and will be excluded from recommendations"
          content:
            application/json:
              schema:
                type: object
                properties:
                  dislikedId:
                    type: integer
                    description: "Unique identifier for the newly created disliked entry"
                    example: 456
                  placeId:
                    type: integer
                    description: "ID of the place that was added to dislikes"
                    example: 4321
                  links:
                    type: object
                    additionalProperties:
                      $ref: "#/components/schemas/HypermediaLink"
                    description: "Available actions for this disliked place"
                    example:
                      self:
                        href: "/users/16180/disliked-places"
                        method: GET
                      remove:
                        href: "/users/16180/disliked-places/456"
                        method: DELETE
                      recommendations:
                        href: "/users/16180/recommendations{?latitude,longitude}"
                        method: GET
                      user-profile:
                        href: "/users/16180/profile"
                        method: GET
        '404':
          description: "User or place not found - Either the user or the specified place does not exist"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: USER_OR_PLACE_NOT_FOUND
                message: "User with ID 12345 or place with ID 4321 not found"
        '409':
          description: "Conflict - Place already in disliked places"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: PLACE_ALREADY_DISLIKED
                message: "Place is already in user's disliked places"
                details:
                  placeId: 4321
                  userId: 16180
  "/users/{userId}/disliked-places/{dislikedId}":
    delete:
      summary: "Remove a place from user's disliked places"
      operationId: removeDislikedPlace
      description: |
        **FR24 - The user should be able to manage places excluded from recommendations.**

        Removes a place from the user's disliked places collection, allowing it to appear in recommendations again.
      tags: ["Favourite & Disliked Places"]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          description: "Unique identifier of the user removing a place from their dislikes"
        - name: dislikedId
          in: path
          required: true
          schema:
            type: integer
          description: "Unique identifier of the disliked entry to be removed from the user's collection"
      responses:
        '204':
          description: "Disliked place successfully removed from user's collection with no content returned"
        '404':
          description: "User or disliked entry not found - Either the user or the specified disliked entry does not exist"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: DISLIKED_NOT_FOUND
                message: "Disliked entry with ID 456 not found for user 12345"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "User authentication using JWT token with userId, email, and role claims"
    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Admin authentication using JWT token. Requires role: admin claim in the JWT payload"
  schemas:
...
    PreferenceCategory:
      type: string
      description: "Possible categories for preferences and place classification"
      enum: [MUSEUM, BEACH, PARK, RESTAURANT, NIGHTLIFE, SHOPPING, SPORTS, CULTURE]
...
    Place:
      type: object
      required: [placeId, name]
      properties:
        placeId:
          type: integer
          description: "Unique identifier for the place"
          example: 4321
        name:
          type: string
          description: "Name of the place"
          example: "White Tower of Thessaloniki"
        description:
          type: string
          description: "Detailed description of the place"
          example: "The White Tower is a historical monument and museum in Thessaloniki, Greece. It is the city's most famous landmark and a symbol of its rich history."
        category:
          $ref: "#/components/schemas/PreferenceCategory"
        location:
          $ref: "#/components/schemas/Location"
          description: "Geographic location of the place with coordinates and address details"
        website:
          type: string
          format: uri
          description: "Official website URL"
          example: "https://www.lpth.gr"
        rating:
          type: number
          format: float
          minimum: 0
          maximum: 5
          description: "Average rating of the place"
          example: 4.7
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
          description: "List of reviews for this place"
        createdAt:
          type: string
          format: date-time
          description: "Place creation timestamp"
    Location:
      type: object
      required: [latitude, longitude]
      properties:
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
          description: "Latitude coordinate in decimal degrees"
          example: 37.9688
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
          description: "Longitude coordinate in decimal degrees"
          example: 23.7281
        address:
          type: string
          description: "Street address of the location"
          example: "Leoforos Nikis"
        city:
          type: string
          description: "City where the location is situated"
          example: Thessaloniki
        country:
          type: string
          description: "Country where the location is situated"
          example: Greece
        postalCode:
          type: string
          description: "Postal/ZIP code of the location"
          example: "54640"
...
    DislikedPlace:
      type: object
      properties:
        dislikedId:
          type: integer
          description: "Unique identifier for the disliked entry"
          example: 456
        placeId:
          type: integer
          description: "ID of the disliked place"
          example: 4321
        place:
          $ref: "#/components/schemas/Place"
    Review:
      type: object
      required: [reviewId, userId, placeId, rating]
      properties:
        reviewId:
          type: integer
          description: "Unique identifier for the review"
          example: 7853
        userId:
          type: integer
          description: "ID of the user who wrote the review"
          example: 16180
        placeId:
          type: integer
          description: "ID of the place being reviewed"
          example: 4321
        rating:
          type: integer
          minimum: 1
          maximum: 5
          description: "Star rating given to the place (1-5 stars)"
          example: 5
        comment:
          type: string
          description: "Optional text comment from the reviewer"
          example: "The White Tower is a must-visit! The view from the top is breathtaking and the museum inside is very informative."
        createdAt:
          type: string
          format: date-time
          description: "Timestamp when the review was created"
          example: "2025-05-15T14:30:00Z"
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: "Machine-readable error code for programmatic handling"
          example: USER_NOT_FOUND
        message:
          type: string
          description: "Human-readable error message describing the issue"
          example: "User with ID 12345 not found"
        details:
          type: object
          description: "Optional additional error details and context"
          additionalProperties: true
          example:
            field: userId
            value: 12345
            timestamp: "2025-06-01T10:30:00Z"
    HypermediaLink:
      type: object
      properties:
        href:
          type: string
          description: "URI of the related resource"
        method:
          type: string
          enum: [GET, POST, PUT, DELETE, PATCH]
          description: "HTTP method for the link"
...